Update to aws-sdk-core v2.0.0.rc16
The API has changed in to significant ways: the way to crete an S3 client is different, and you now need to send the content MD5 as a Base64 encoded string instead of as a hexadecimal string. It’s unclear whether or not the content MD5 was even sent before, since the Base64 format is mandated by the S3 API and not aws-sdk-core.