Make it possible to run a command before the artifact is built
It will be run in the directory with the cleanly checked out code. The idea is that if you have things that need to be built from your source this is when you can do it.

For Ruby projects I think it would make sense to include a `:build_block` option (or some other name) that let you run some Ruby code instead of having to shell out.