Allow custom app_name when creating artifact
Currently it’s only possible to build an artifact that uses the same
name as the basename of the project, so it’s essentially impossible to
build more than one artifact per Git repository.

This commit adds an `:app_name` option that will only change the
`app_name` part of the artifact name. It could be debated whether this
option should be passed down to the #create method of the archiver as
well, but for simplicity it’s currently not.