Recode S3 metadata using ASCII \u{CODEPOINT} sequences
The motivation for doing this is to ensure that people with non-ASCII names in their Git configuration is able to release artifacts to S3.

With the current state of affairs, the release fails with a rather cryptic Aws::S3::Errors::SignatureDoesNotMatch. This turns out to be because AWS changes any non-ASCII bytes to a literal question mark `?`, when creating the canonical request. The documentation says nothing about this, so perhaps it should not come as a surprise that Amazon's own tooling doesn't do this.

By fixing the signature calculation, putting UTF-8 data to S3 actually works, but when retrieving the data it is returned in MIME Encoded-Word format, which is typically not handled by HTTP clients (awscli being an important example).

The proposed encoding is a bit easier on the eyes than the MIME option, for clients with no support for encoded headers, and where the value is mostly ASCII.